pyproject.toml
[tool.poetry]
name = "exchange-latency-explorer"
version = "0.1.0"
description = ""
authors = ["rvdw <rohan.vanderwalt@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.12.3"
paramiko = "^3.5.0"
pyyaml = "^6.0.2"
ccxt = "^4.4.31"
pandas = "^2.2.3"
plotly = "^5.24.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
